+++ app/src/main/java/com/example/spartube/comment/BottomSheetCommentFragment.kt
@@ -1,4 +1,4 @@
-package com.example.spartube.shorts
+package com.example.spartube.comment
 
 import android.app.Dialog
 import android.os.Bundle
@@ -13,7 +13,6 @@ import com.example.spartube.data.service.RetrofitModule
 import com.example.spartube.databinding.ShortsCommentBinding
 import com.example.spartube.shorts.adapter.CommentBindingModel
 import com.example.spartube.shorts.adapter.CommentSetBindingModel
-import com.example.spartube.comment.CommentsAdapter
 import com.example.spartube.comment.reply.ReplyAdapter
 import com.example.spartube.util.Converter
 import com.google.android.material.bottomsheet.BottomSheetBehavior
