+++ app/src/main/java/com/example/spartube/comment/reply/ReplyAdapter.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.spartube.databinding.ShortsPageCommentItemBinding
 import com.example.spartube.databinding.ShortsPageReplyitemBinding
-import com.example.spartube.shorts.ViewType
+import com.example.spartube.comment.ViewType
 import com.example.spartube.shorts.adapter.CommentBindingModel
 import java.lang.RuntimeException
