+++ app/src/main/java/com/example/spartube/detail/DetailPageFragment.kt
@@ -13,8 +13,8 @@ import androidx.viewpager2.widget.ViewPager2
 import com.example.spartube.R
 import com.example.spartube.data.service.RetrofitModule
 import com.example.spartube.databinding.FragmentDetailPageBinding
-import com.example.spartube.shorts.BottomSheetCommentFragment
-import com.example.spartube.shorts.ViewType
+import com.example.spartube.comment.BottomSheetCommentFragment
+import com.example.spartube.comment.ViewType
 import com.example.spartube.shorts.adapter.CommentSetBindingModel
 import com.example.spartube.db.AppDatabase
 import com.example.spartube.db.MyPageEntity
